Crate wit_compiler

Re-exports

pub use wit_syntax as ast;

Modules

access

Indirection used when navigating the crate::hir and crate::queries::metadata.

diagnostics

Errors and user-facing messages that may be generated as the result of analysis.

hir

The high-level intermediate representation.

queries

Various operations used by the compiler and language server.

traverse

Utilities for traversing a tree_sitter::Tree.

Structs

Compiler

The WAI language’s compiler frontend.

Jar

A salsa::jar::Jar containing all queries and tracked types used by this crate.

Text

A cheaply cloneable, immutable string.

Tree

A wrapper around tree_sitter::Tree that is comparable.

Traits

Db

The trait that all salsa::Database implementations must implement to be used with this crate’s queries.