Module wit_compiler::diagnostics

Errors and user-facing messages that may be generated as the result of analysis.

Structs

Bug

Deprecated

DiagnosticInfo

Diagnostics

An accumulator for all Diagnostics that have been emitted.

DuplicateName

LintContext

LintPass

Location

The location of an element within the workspace.

MismatchedPackageDeclaration

MultipleConstructors

MultiplePackageDocs

SyntaxError

UnknownName

The user referenced an unknown identifier.

UnknownPackage

The user referenced an unknown identifier.

check_workspace

Enums

Diagnostic

Diagnostic messages that are emitted using Diagnostics.

Constants

LINT_PASSES

Traits

IntoDiagnostic

Functions

all_diagnostics

Information about all know Diagnostic types.

check_all

Check an entire workspace.

check_workspace