Trait wit_compiler::diagnostics::IntoDiagnostic

pub trait IntoDiagnostic: Into<Diagnostic> {
    const CODE: &'static str;
    const MESSAGE: &'static str;
    const VERBOSE_DESCRIPTION: &'static str;
    const SEVERITY: Severity;

    // Required method
    fn update_diag(&self, diag: Diagnostic<FilePath>) -> Diagnostic<FilePath>;

    // Provided method
    fn as_diagnostic(&self) -> Diagnostic<FilePath> { ... }
}

Required Associated Constants

const CODE: &'static str

A unique code which can be used when referring to this error.

const MESSAGE: &'static str

A simple message that is displayed with the error.

const VERBOSE_DESCRIPTION: &'static str

A verbose explanation of the error.

const SEVERITY: Severity

The default severity.

Required Methods

fn update_diag(&self, diag: Diagnostic<FilePath>) -> Diagnostic<FilePath>

Provided Methods

fn as_diagnostic(&self) -> Diagnostic<FilePath>

Object Safety

This trait is not object safe.