Module wit_compiler::queries

Various operations used by the compiler and language server.

Re-exports

pub use self::metadata::file_items;

pub use self::metadata::file_items;

pub use self::metadata::package_items;

pub use self::metadata::package_items;

Modules

metadata

Structs

Ast

FilePath

The path to a SourceFile in the Workspace.

LineNumbers

A lookup table for efficiently looking up line numbers in a file.

Package

A group of files in a Workspace.

PackageId

SourceFile

A file attached to a Workspace.

Workspace

A workspace keeps track of all files known to the compiler.

WorkspaceFiles

An implementation of codespan_reporting::files::Files which will query the Db as required.

calculate_line_numbers

imported_types

lower_package

parse

selection_ranges

workspace_packages

Functions

calculate_line_numbers

imported_types

Determine the items that have been imported inside a particular world or interface.

lower_package

parse

Parse a SourceFile into an Ast, emitting any SyntaxErrors as they occur.

selection_ranges

Calculate successively wider ranges of tokens.

top_level_names

Find all the names accessible at the top level of a file.

workspace_packages