Trait wit_compiler::traverse::Cursor

pub trait Cursor {
    type Node;

    // Required methods
    fn goto_first_child(&mut self) -> bool;
    fn goto_parent(&mut self) -> bool;
    fn goto_next_sibling(&mut self) -> bool;
    fn node(&self) -> Self::Node;
}

Trait which represents a stateful cursor in a n-ary tree. The cursor can be moved between nodes in the tree by the given methods, and the node which the cursor is currently pointing at can be read as well.

Required Associated Types

type Node

The type of the nodes which the cursor points at; the cursor is always pointing at exactly one of this type.

Required Methods

fn goto_first_child(&mut self) -> bool

Move this cursor to the first child of its current node.

This returns true if the cursor successfully moved, and returns false if there were no children.

fn goto_parent(&mut self) -> bool

Move this cursor to the parent of its current node.

This returns true if the cursor successfully moved, and returns false if there was no parent node (the cursor was already on the root node).

fn goto_next_sibling(&mut self) -> bool

Move this cursor to the next sibling of its current node.

This returns true if the cursor successfully moved, and returns false if there was no next sibling node.

fn node(&self) -> Self::Node

Get the node which the cursor is currently pointing at.